requirements.txt
arrow==0.15.4
cdsapi==0.2.3
geopy==1.20.0
matplotlib==3.1.1
MetPy==0.12.0
networkx==2.4
numpy==1.21.6
pandas==0.25.3
Pillow==6.2.1
pmdarima==1.8.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.22.2
scipy==1.3.2
tqdm==4.38.0
xarray==0.14.0
bresenham==0.2.1